Revert "INC-1165 Performance improvements for Case  Note type count (#1538)"
This reverts commit 1c0cc23af858d00142ba82ef2c7201f1aaf1d437.

+++ src/main/java/uk/gov/justice/hmpps/prison/api/model/CaseNoteTypeSummaryRequest.java
@@ -39,7 +39,7 @@ public class CaseNoteTypeSummaryRequest {
     public static class BookingFromDatePair {
 
         @Schema(description = "Booking Id")
-        private Long bookingId;
+        private Integer bookingId;
 
         @Schema(description = "Only case notes occurring on or after this date (in YYYY-MM-DDTHH:MM:SS format) will be considered.", example = "2018-11-01T12:00:00")
         private LocalDateTime fromDate;
